Patch Vec::with_capacity & Vec::push to use Boehm allocator
A common GC pattern is to pass a `Vec<T>` to a `Gc` to be managed by the
collector. This requires the collector to register a finalizer in order
to call `Vec`'s drop method - even if `needs_drop::<T>() == false`! This
commit seeks to fix that by allowing the compiler to ignore `Vec<T>::drop`
(and hence, remove the need to register a finalizer) if it is safe to do
so.

This is possible by doing three things:

	* A new trait `ManageableContents` is introduced in `libcore`,
	  this is to be implemented on every `T` whose contents are safe
	  to be allocated with Boehm and not be dropped.
	  (see commit 197f42)

	* Stubs for the Boehm allocator are added to `liballoc` to allow
	  the compiler to switch between Global and Boehm allocation
	  (see commit a5c61f):

	* In this commit, a MIR transform is added which searches for
	  `Vec<T>::new()` constructors in user code. When found, it
  	  checks if `T: ManageableContents` and, if so, will patch the
	  constructor to create a `RawVec` with a `Boehm` backing
	  allocator.