Add mem::needs_finalizer::<T>() intrinsic
This method is identical to mem::needs_drop::<T>() with one important
difference: it returns `false` if `T` implements `NoFinalize`. For
example:

    struct A;
    impl NoFinalize for A {}

    impl Drop for A {
        fn drop(&mut self) { ... }
    }

    struct B(A);

Here, `needs_finalizer::<A>()` returns `false`. This is because the
implementation of `NoFinalize` on `A` has the effect of "cancelling out"
the `Drop` impl. In other words, `needs_finalizer` treats the
`NoFinalize` trait as having a higher precedence than the `Drop` trait.

Calling `needs_finalizer::<B>() also returns `false`. This is because 1)
`B` has no destructor, and 2) its component type, `A` doesn't need
finalizing.

However, if we introduce another type:

    struct C(A, String);

Then calling `needs_finalizer::<C>()` will return `true`, due to the
component type, `String`, having a destructor. This is because
`needs_finalizer::<T>` will recurse through each component type of `T`
until it finds a component type with a destructor (which isn't marked
`NoFinalize`), or it has exhausted the entire type.