Add intrinsic for querying type's GC layout
For precise object marking, it's necessary for the collector to know
which fields could lead to pointers, and which it can safely ignore.
This has two benefits: 1) it speeds up marking, and 2) it can prevent
memory leaks by disambiguating non-obvious non-pointers.

The core::mem::gc_layout::<T>() returns a pair of u64s: the first value
is a bitmap representing which words in the type's layout are pointers;
the second value is the size of the bitmap, (e.g. a size of 3 indicates
only the last 3 bits in the bitmap are of interest).

Currently the bitmap will treat any field in a type which is at least 1
word in size as a potential pointer value. To opt out of this for
specific fields, the #[no_trace] annotation can be used. Consider the
struct below:

    struct S {
        a: usize,
        #[no_trace]
        b: usize,
    }

Calling `gc_layout::<S>()` would return `(0xFFFFFFFD, 2)`. Fields in the
bitmap grow upwards from the least significant bit. In this example,
only the last hexadecimal digit (D) is of interest.

     1111 1111 1111 1111 1111 1111 1111 1101
                                          ^.. size (2)

It states that only the first word _might_ be a pointer. By default,
unused fields in this bitmap are set to 1. In this example, every bit
except the last two are unused because the size returned was 2.

This intrinsic is evaluated after type layouts have been established, so
the bitmap will be correct even if struct fields are shuffled around by
the optimiser. It is not guaranteed to match the source order.

Structs which have an alignment smaller than a machine word can be
completely ignored by the collector, since they could never contain
fields large enough to hold pointers. For example, the following struct:

    struct T {
        a: u32,
        b: u16,
        c: u32,
        d: u16
    }

Would have the this bitmap: `0x00000000`.

This intrinsic is a first draft and currently only works on structs.