Automatically exclude certain types from GC marking
This commit introduces the `NoTrace` auto trait. An auto trait is a
marker trait that is applied to every type unless it's explicitly opted
out (e.g. `Send` and `Sync`). Certain types that are used to hold Gc
pointers (e.g. `*mut T`, `*const T`) are _negatively implemented_ for
`NoTrace`. When building a type bitmap, `gc_layout` will now also check
if a field implements `NoTrace`.

The power of auto traits means that this also works transitively, so
GC pointers buried deep in nested type layouts will not be missed.